requirements.txt
pyserial
flask
pyyaml
Flask-MySQLdb
mysql-connector
DateTime
datetime
